requirements.txt
numpy
matplotlib
pyyaml
scipy
torch
transformers
pillow
opencv-python
tqdm
scikit-learn
scikit-image
pandas
